backend/requirements.txt
# Core Django framework
Django==4.2.7
djangorestframework==3.14.0

# AWS SDK for Python
boto3==1.34.10
botocore==1.34.10

# Environment variables management
python-dotenv==1.0.0

# CORS headers for frontend communication
django-cors-headers==4.3.1

# JWT authentication
PyJWT==2.8.0

# Password hashing
bcrypt==4.1.2

# HTTP requests
requests==2.31.0

# Date/time utilities
python-dateutil==2.8.2

# UUID generation (built-in but listed for clarity)
# uuid is part of Python standard library

# Development server
gunicorn==21.2.0

# AWS Lambda deployment
zappa==0.58.0
